refactor(sandbox): consolidate cleanup contracts
- Share receipt, response, path and journal validation.
- Exercise the production cleanup coordinator in tests.
