+++ lib/sandbox/control/server.ts
@@ -1,3 +1,4 @@
+import { finalizationTerminalResponse } from './finalization-response.ts';
 import { completedReentryView } from './completed-reentry.ts';
 import fs from 'node:fs';
 import path from 'node:path';
@@ -212,33 +213,7 @@ function finalizationRecoveryResponse(
   } catch {
     return { status: 'deferred' };
   }
-  const pendingSteps = [
-    receipt.taskComment === 'pending' ? 'task-comment' : null,
-    receipt.verification === 'pending' ? 'verification' : null
-  ].filter((step): step is string => step !== null);
-  const completedSteps = ['lifecycle', receipt.taskComment === 'pending' ? null : 'task-comment', receipt.verification === 'pending' ? null : 'verification']
-    .filter((step): step is string => step !== null);
-  const warnings = receipt.warnings
-    .filter((warning) => warning.status === 'open')
-    .map(({ status: _status, resolvedAt: _resolvedAt, ...warning }) => warning);
-  const result = {
-    status: 'completed', changed: false, taskId: manifest.taskId,
-    lifecycle: { status: 'no-op', changed: false, error: null },
-    taskComment: receipt.taskComment === 'pending' ? null : { status: 'no-op', changed: false, error: null },
-    verification: receipt.verification === 'pending' ? null : { status: 'no-op', changed: false, error: null },
-    completedSteps, pendingSteps,
-    result: pendingSteps.length > 0 || receipt.warningProjection === 'pending' || warnings.length > 0
-      ? 'completed_with_warnings' : 'completed',
-    warnings, error: null
-  };
-  return {
-    status: 'matched',
-    response: {
-      version: 2, id: requestId, phase: 'completed', exitCode: 0,
-      stdout: `${JSON.stringify({ version: 1, status: 'completed', changed: false, accepted: true, result, error: null })}\n`,
-      stderr: '', error: null
-    }
-  };
+  return { status: 'matched', response: finalizationTerminalResponse(manifest.taskId, requestId, receipt) };
 }
 
 function publishFinalizationTaskView(
